packages/python/otlp-stdout-span-exporter/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "otlp-stdout-span-exporter"
version = "0.1.2"
description = "OpenTelemetry span exporter that writes to stdout in OTLP format"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Apache-2.0" }
authors = [
    { name = "Alessandro Bologna", email = "alessandro.bologna@gmail.com" }
]
keywords = ["opentelemetry", "tracing", "spans", "otlp", "stdout", "exporter"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Monitoring",
]
dependencies = [
    "opentelemetry-sdk>=1.29.0",
    "opentelemetry-exporter-otlp-proto-common>=1.29.0",
]

[project.urls]
homepage = "https://github.com/dev7a/serverless-otlp-forwarder/"
repository = "https://github.com/dev7a/serverless-otlp-forwarder/tree/main/packages/python/otlp-stdout-span-exporter"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[tool.hatch.build]
only-packages = true
sources = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]

[tool.ruff.lint.isort]
known-first-party = ["otlp_stdout_span_exporter"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=otlp_stdout_span_exporter --cov-report=term-missing"
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true 